requirements.txt
appdirs==1.4.3
asciitree==0.3.3
astroid==2.3.3
attrs==19.3.0
backcall==0.1.0
certifi==2019.9.11
chardet==3.0.4
ConfigArgParse==0.15.1
datrie==0.8
decorator==4.4.1
docutils==0.15.2
fasteners==0.15
fire==0.2.1
gitdb2==2.0.6
GitPython==3.0.5
idna==2.8
importlib-metadata==0.23
ipython==7.9.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.15.1
jsonschema==3.1.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
monotonic==1.5
more-itertools==7.2.0
numcodecs==0.6.4
numpy==1.17.4
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
pkg-resources==0.0.0
prompt-toolkit==2.0.10
psutil==5.6.5
ptyprocess==0.6.0
Pygments==2.4.2
pylint==2.4.4
pyrsistent==0.15.5
python-dotenv==0.10.3
PyYAML==5.1.2
ratelimiter==1.2.0.post0
requests==2.22.0
six==1.13.0
smmap2==2.0.5
snakemake==5.7.4
termcolor==1.1.0
tqdm==4.38.0
traitlets==4.3.3
typed-ast==1.4.0
urllib3==1.25.7
wcwidth==0.1.7
wrapt==1.11.2
zarr==2.3.2
zipp==0.6.0
